Move repo blobs functionality into service
