+++ packages/pds/src/api/com/atproto/blob.ts
@@ -1,17 +1,14 @@
 import { Server } from '../../../lexicon'
 import * as locals from '../../../locals'
-import * as repo from '../../../repo'
 import ServerAuth from '../../../auth'
 
 export default function (server: Server) {
   server.com.atproto.blob.upload({
     auth: ServerAuth.verifier,
     handler: async ({ input, res }) => {
-      const { db, blobstore } = locals.get(res)
+      const { services } = locals.get(res)
 
-      const cid = await repo.addUntetheredBlob(
-        db,
-        blobstore,
+      const cid = await services.repo.blobs.addUntetheredBlob(
         input.encoding,
         input.body,
       )
