+++ packages/pds/src/api/com/atproto/repo.ts
@@ -101,7 +101,7 @@ export default function (server: Server) {
     handler: async ({ input, auth, res }) => {
       const tx = input.body
       const { did, validate } = tx
-      const { db, blobstore, services } = locals.get(res)
+      const { db, services } = locals.get(res)
       const requester = auth.credentials.did
       const authorized = await services.repo.isUserControlledRepo(
         did,
@@ -153,7 +153,7 @@ export default function (server: Server) {
       await db.transaction(async (dbTxn) => {
         const now = new Date().toISOString()
         const repoTxn = services.repo.using(dbTxn)
-        await repoTxn.processWrites(did, authStore, blobstore, writes, now)
+        await repoTxn.processWrites(did, authStore, writes, now)
       })
     },
   })
@@ -164,7 +164,7 @@ export default function (server: Server) {
       const { did, collection, record } = input.body
       const validate =
         typeof input.body.validate === 'boolean' ? input.body.validate : true
-      const { db, blobstore, services } = locals.get(res)
+      const { db, services } = locals.get(res)
       const requester = auth.credentials.did
       const authorized = await services.repo.isUserControlledRepo(
         did,
@@ -209,7 +209,7 @@ export default function (server: Server) {
 
       await db.transaction(async (dbTxn) => {
         const repoTxn = services.repo.using(dbTxn)
-        await repoTxn.processWrites(did, authStore, blobstore, [write], now)
+        await repoTxn.processWrites(did, authStore, [write], now)
       })
 
       return {
@@ -227,7 +227,7 @@ export default function (server: Server) {
     auth: ServerAuth.verifier,
     handler: async ({ input, auth, res }) => {
       const { did, collection, rkey } = input.body
-      const { db, services, blobstore } = locals.get(res)
+      const { db, services } = locals.get(res)
       const requester = auth.credentials.did
       const authorized = await services.repo.isUserControlledRepo(
         did,
@@ -248,7 +248,7 @@ export default function (server: Server) {
 
       await db.transaction(async (dbTxn) => {
         const repoTxn = services.repo.using(dbTxn)
-        await repoTxn.processWrites(did, authStore, blobstore, write, now)
+        await repoTxn.processWrites(did, authStore, write, now)
       })
     },
   })
