+++ packages/pds/src/repo/index.ts
@@ -1,3 +1,2 @@
-export * from './blobs'
 export * from './prepare'
 export * from './types'
