+++ packages/pds/src/services/index.ts
@@ -1,3 +1,4 @@
+import { BlobStore } from '@atproto/repo'
 import Database from '../db'
 import { MessageQueue } from '../stream/types'
 import { ActorService } from './actor'
@@ -7,10 +8,11 @@ import { RepoService } from './repo'
 export function createServices(
   db: Database,
   messageQueue: MessageQueue,
+  blobstore: BlobStore,
 ): Services {
   const actor = new ActorService(db)
   const record = new RecordService(db, messageQueue)
-  const repo = new RepoService(db, messageQueue)
+  const repo = new RepoService(db, messageQueue, blobstore)
   return { actor, record, repo }
 }
