+++ packages/pds/src/stream/consumers/index.ts
@@ -7,9 +7,11 @@ import RemoveUpvoteConsumer from './remove-upvote'
 import CreateNotificationConsumer from './create-notification'
 import DeleteNotificationsConsumer from './delete-notifications'
 import { MessageQueue } from '../types'
+import { BlobStore } from '@atproto/repo'
 
 export const listen = (
   messageQueue: MessageQueue,
+  blobstore: BlobStore,
   auth: ServerAuth,
   keypair: DidableKey,
 ) => {
@@ -18,7 +20,10 @@ export const listen = (
   }
   messageQueue.listen('add_member', new AddMemberConsumer())
   messageQueue.listen('remove_member', new RemoveMemberConsumer())
-  messageQueue.listen('add_upvote', new AddUpvoteConsumer(getAuthStore))
+  messageQueue.listen(
+    'add_upvote',
+    new AddUpvoteConsumer(getAuthStore, messageQueue, blobstore),
+  )
   messageQueue.listen('remove_upvote', new RemoveUpvoteConsumer())
   messageQueue.listen('create_notification', new CreateNotificationConsumer())
   messageQueue.listen('delete_notifications', new DeleteNotificationsConsumer())
